lint(client): ignore build outputs; fix no-explicit-any, prefer-const, and empty interface issues; split components to satisfy react-refresh; add lint CI workflow
