+++ apps/client/src/components/lexical/ImageNode.tsx
@@ -9,8 +9,9 @@ import type {
   Spread,
 } from "lexical";
 
-import { $applyNodeReplacement, DecoratorNode } from "lexical";
+import { $applyNodeReplacement, DecoratorNode, type LexicalEditor } from "lexical";
 import * as React from "react";
+import ImageComponent from "./ImageComponent";
 
 export interface ImagePayload {
   altText: string;
@@ -131,35 +132,9 @@ export class ImageNode extends DecoratorNode<React.JSX.Element> {
     return this.__fileName;
   }
 
-  decorate(_editor: any, _config: EditorConfig): React.JSX.Element {
+  decorate(_editor: LexicalEditor, _config: EditorConfig): React.JSX.Element {
     return <ImageComponent src={this.__src} altText={this.__altText} fileName={this.__fileName} />;
   }
-}
-
-function ImageComponent({
-  src,
-  altText,
-  fileName,
-}: {
-  src: string;
-  altText: string;
-  fileName?: string;
-}) {
-  return (
-    <div className="relative inline-block my-2">
-      <img
-        src={src}
-        alt={altText}
-        className="max-w-full h-auto rounded-lg border border-neutral-200 dark:border-neutral-700"
-        style={{ maxHeight: "300px" }}
-      />
-      {fileName && (
-        <div className="absolute bottom-0 left-0 right-0 bg-black/50 text-white text-xs px-2 py-1 rounded-b-lg">
-          {fileName}
-        </div>
-      )}
-    </div>
-  );
 }
 
 export function $createImageNode({
